requirements.txt
# ============================================
# MediChat AI - Backend Requirements
# ============================================
# Medical AI Chatbot with Voice & Image Support
# Python 3.9+
# ============================================

# Core Framework
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1

# Environment & Configuration
python-dotenv==1.0.0

# AI & Machine Learning
groq==0.15.0                    # AI model integration
pillow==11.1.0                  # Image processing

# Text-to-Speech (TTS)
elevenlabs==1.50.3              # Premium TTS service
gtts==2.5.4                     # Google Text-to-Speech

# Speech Recognition
speechrecognition==3.13.0       # Voice input processing
pydub==0.25.1                   # Audio file manipulation

# HTTP & API Clients
requests==2.32.3                # HTTP library
httpx==0.28.1                   # Async HTTP client
certifi==2024.12.14             # SSL certificates

# Data Validation
pydantic==2.10.5                # Data validation & settings

# ============================================
# Installation Instructions:
# pip install -r requirements.txt
# ============================================
